+++ Changelog.md
@@ -3,6 +3,7 @@ OpenCore Changelog
 #### v0.6.1
 - Improved recognition of early pressed hotkeys, thx @varahash
 - Made DMG loading support configurable via `DmgLoading`
+- Added iMac20,1 model code
 
 #### v0.6.0
 - Fixed sound corruption with AudioDxe